Making a better test case
