+++ tests/Functional/IntrospectionTest.php
@@ -5,6 +5,7 @@ namespace Digia\GraphQL\Test\Functional;
 use function Digia\GraphQL\buildSchema;
 use Digia\GraphQL\Test\TestCase;
 use function Digia\GraphQL\graphql;
+use function Digia\GraphQL\Type\newScalarType;
 
 class IntrospectionTest extends TestCase
 {
@@ -451,35 +452,4 @@ class IntrospectionTest extends TestCase
         ], $result);
     }
 
-    public function testCanIntrospectNameOnADateScalar()
-    {
-        $schema = '
-        scalar Date
-        type Query {
-            hello: String
-        }
-        ';
-        $schema = buildSchema($schema);
-
-        $query = '
-        query IntrospectionDroidDescriptionQuery {
-          __type(name: "Date") {
-            name
-          }
-        }
-        ';
-
-        $result = graphql($schema, $query);
-
-        $this->assertNull($result['errors']);
-        $this->assertEquals([
-            'data' => [
-                '__type' => [
-                    'name'        => 'Date',
-                ],
-            ],
-        ], $result);
-
-
-    }
 }
